// models/ui_menu.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery'

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type UiMenu = IContentItem<{
	subtitle: Elements.TextElement
	subitems: Elements.LinkedItemsElement<IContentItem>
	item: Elements.LinkedItemsElement<IContentItem>
	title: Elements.TextElement
}>